lxd/storage/drivers/powerflex: Copy the API request body
In case the API's access token expires a new token is generated and the request
is tried one more time.
As the body is provided as a reader, this reader will be empty when issuing the second request
as its contents were already read in the request from before.
Prevent this by creating a copy of the readers contents and provide a fresh reader from the copy
for every new request.

This was causing 400 Bad Requests in some cases when the token expired and the request which
is going to be retried had something initially set in it's request body.
For requests without anything in the body this was still working fine and therefore stayed unnoticed until now.
It first occurred in a very slow PowerFlex environment in which an instance migration took longer
than the token lifetime of around 10mins.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>